refactor: 카드 화면 관련 코드 개선
- CardsActivity에서 CardsScreen 분리
